+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Tooltip: fix children not having consistent ref on re-render (like when updating the tooltip label)
+
 ## [3.7.1][] - 2024-05-21
 
 ### Fixed